usbboot: Add a "-f" option for fastboot
This patch modifies the Makefile to generate secondstage.c,
which defines iboot_size and the iboot_data structure.
This file is compiled and linked with usbboot.

We also modify the tools/usbboot.c to recognize the command
line option of "usbboot -f". The -f option will indicate
that the user wants to enter into FASTBOOT mode and will prevent
the need of providing a 3rd parameter to usbboot.

command:
sudo ./out/panda/usbboot -f

Change-Id: I6b4baa26876a35529a5edc3ca2f61cca5881c739
Signed-off-by: Regis Goupil <x176713@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>