fastboot: Issue with get_userdata_size
The partition names are being stored in unicode
format based on what the kernel expects it to be.
We were trying to look for a partition name in unicode
format by doing a strcmp. This is incorrect, fixed
with this patch.

This patch adds an api to convert the partition name
from unicode to non-unicode format.

Change-Id: Iaff163bb2d102e63d9697176f3a7ef7a518e7d0c
Signed-off-by: Regis Goupil <x176713@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>