mkheader: Use the correct arg for adding GP hdr
The return value of the command 'wc -c filename' is size of filename
followed by the filename. This would mean that argv[2] will contain
the size of the filename and argv[3] will contain the filename.
The string "no_gp_hdr" or "add_gp_hdr" will then be contained in
argv[4].

With the above change, we also need to modify the Makefile to define
TEXT_BASE else we FAIL the compilation with a Segmentation fault.

Change-Id: Iad5f3b69dad595cc70e42e795aa27e4f22edd2a3
Signed-off-by: Regis Goupil <x176713@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>