Make: Ensure the right lds file is used for linking
Currently due to error in target generation, the wrong
eboot.lds linker is used for all builds instead of the right
linker file. so fix that - as part of this, iboot now should
explicitly state what linker file it uses as well.. so copy
the eboot linker file for iboot as a start point.

Change-Id: I6d8cd5b30b12921107297c9d3d84ec93aabed37a
Signed-off-by: Nishanth Menon <nm@ti.com>