usb: linux: Remove ZLP for data size >64KiB
Sending a ZLP after 64KiB for images >64KiB results in
SEC_RESET_CERT_ISW in PPA. Believed that ZLP was necessary
due to some limitation coming from the buffer size getting used on
target side and ROM code requiring ZLP for buffer end condition. This
assumption was not tested at time of introduction.

Removing this ZLP results in correct operation of peripheral
booting for all validly sized images (<120KiB).

ZLP is still necessary for any image that is a multiple of 512
packets. ES2.0 will remove the need for any ZLP.

Change-Id: Ide1057675976ee164fc5e7809d77aeb605fe0e77
Signed-off-by: Bryan Buckley <bryan.buckley@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>