omap: usb-rom-api: Initialized the rom usb for memory boot
When we boot from a memory device like eMMC, there is no existing
USB connection to open. The rom usb has to be initialized before
we can use it. The device will not enumerate without the init
being called.

We need to close the usb before we boot the kernel else we may
see issues with the USB when we try to enable it during kernel
boot.

Change-Id: Ib13cd1623ae4f0f3c2a72326eda38bba4e962e0f
Signed-off-by: Olivier Deprez <o-deprez@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>