omap: don't use hardcoded values for memory size
Define a start and size for each hardware target
in their config file.
Use MEMORY_SIZE and MEMORY_BASE values when setting
the atag to pass to the kernel.
Use MEMORY_BASE as reference for memory configurations
This allows to change the memory size passed to the
kernel as requested by program management limiting
the memory amount seen by the kernel.
Tested configurations are limited to 1 GB and 2 GB
for Omap5 and 1 GB for Omap4.

Change-Id: I2c6e02e4c73721c253da1b1bcd9b6f1e2dd2a082
Signed-off-by: Regis Goupil <x0176713@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>