fastboot: use the mem alloc to allocate memory for data download
We were using hardcoded addresses in SDRAM to download the data
for flashing/booting. This patch removes that and adds functionality
to request for memory in SDRAM based on the size of the image that
needs to be downloaded. Once the flashing is done, we free up that
memory by marking it as unused.

Change-Id: Id6627dc1524bd8e0746848129f5b4442d6bceb9d
Signed-off-by: Christina Warren <cawarren@ti.com>