omap5evm: Fix wrong pull direction of fref_clk_req0 line
In OMAP5 sEVM/uEVM configurations, we use crystal + OMAP internal
oscillator. In this configuration, FREF_CLK_IOREQ is an input
controlling auxclk0 output. If we do not retain the reset pull
DOWN for the pad, this results in OMAP thinking that it needs
to supply sysclk as output in auxclk0, this results in sysclk
never idling and preventing device wide OFF state with sys_pwr_req
being asserted in OFF mode. Even though this will be fixed
in kernel and make kernel independent of the bootloader, it
is appropriate meanwhile to ensure that the bootloader
does the "right thing" in ensuring appropriate PULL is set
corresponding to the platform.

Change-Id: Ieeec5586548247324b3ef400332185a30d5bec10
Reported-by: Oleksandr Kozaruk <oleksandr.kozaruk@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>