mem_alloc: Fix/reimplement alloc_memory
Fix and re-implement memory management because
of following:

1. In case where there is a previosly freed memory
section in the heap chain and if that section
is larger than the size of memory being requested,
there are following issues:

a. Current pointer manipulation mechanism looses
references to all the nodes following free section.

b. Pointer manipulation to compute the address
to the next node is also wrong
next += sizeof(head) + (size * sizeof(u32));
instead of
next += sizeof(head) + size
since the "size" variable is already in bytes.

2. Concept of available memory is also incorrect.
We are merely adding up all the fragmented memories
available to compute free memory. But yet, we have
no way of handling fragmented memory on a per-request
basis.

3. Allocated memory also has to be aligned to 512 bytes
as some of the buffers might be given to MMC/SATA devices

Change-Id: Ib844ae1d117484d316d0083856e2279c0c6eee5b
Signed-off-by: Viswanath Puttagunta <vishp@ti.com>