omap5: clocks: keep emif in sw mode
OMAP5432 requires emif to be always in Software Supervised mode.
So disable HW_AUTO for CM_MEMIF_CLKSTCTRL at bootloader level.

For Kernel:
For OMAP5430, kernel will take control and put the EMIF back to
HW_AUTO

For OMAP5432, change will be done to not allow HW_AUTO for emif

Change-Id: Ib753b1e02a32b7422f854d8888f0a2cf4d96098c
Signed-off-by: Olivier Deprez <o-deprez@ti.com>
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>