Two stage: Compile time memory tests inclusion
This patch demonstrate how to include a memory test
during the first stage of the boot.

The memory tests can be run either during memory
boot or peripheral boot depending upon 2 #define
in the configuration.

The memory tests included, taken from aboot, are
just intended to be a quick very simple example.

Change-Id: I7df337cdbcd867bac360f97b56f46b4cefbde025
Signed-off-by: Regis Goupil <x0176713@ti.com>