Two stage: Dynamic Memory Test during first stage
Passing a parameter -m (or -M) to usbboot will
trigger execution of Memory Tests during first
boot.
Note that #define created during previous patch
are required. If defined, the usbboot parameter will
be taken in account, if not defined, the parameter
will be ignored.

Memory tests could be run during peripheral boot
by passing this -m option to usbboot but also
systematically during memory boot by setting
the proper define in the config file.
This define is commented in this patch.

Change-Id: Ibd6e6ca1cb513c274bf6b353a0ee088a92859bf2
Signed-off-by: Regis Goupil <x0176713@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>