Two Stage: workaround to fix fasboot in second stage
rom_usb seems unable to handle to handle usb
operations when initaited from sdram.

This patch adds an inderection for these
operations to be handle by the code in sram
even when initiated from sdram.

Until the issue is worked out, this
Worakaround or hack allows fastboot commands
to be processed in the second stage bootloader.

Imprtant note:
Even though usb_open indirection is included
in sboot to satisfy the compiler, it cannot be
called from sboot. The call will fail.

Change-Id: Ia78620f757594d9cc289582a8e02bfacae83f88d
Signed-off-by: Regis Goupil <x0176713@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>