omap: rom_usb: Fix fastboot issue when using mem preferred boot
This patch cleans up the implementation of rom usb_open() to
perform a usb_init() depending on the type of boot that occurs
i.e. memory or peripheral. This change prevents the need for a
usb_init() to be called separately in the code.

This patch fixes issues seen when trying to run fastboot commands
when a memory preferred boot order like eMMC first and USB second
is used.

Change-Id: I400a3eab39ee4c4397c3e13376bb30ebc0d2a0f9
Signed-off-by: Olivier Deprez <o-deprez@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>
Signed-off-by: Regis Goupil <x0176713@ti.com>