omap5: Setting up stack as required by ES2.0
For omap5 ES1.0, the bootloader relies on the
rom-code to setup the stack. This is not the
case anymore for omap5 ES2.0 where the sp
register (r13) maybe undefined.

This patch sets up the omap5 SP register to 0x4031c000
which is compatible with omap5 ES1.0 and omap5 ES2.0.

This patch also sets up the omap4 SP register to 0x4030c000
which is compatible with 4430, 4460 and 4470.

A special note about SATA support on omap5 ES1.0:
area [ 0x4031d000 - 0x4031ffff ] should be avoided.
Since SATA is not fully supported with ES1.0,
(SATA boot not supported) this shouldn't be an issue.

If a bigger stack is required during the second stage,
where code is most likely to be added, change the SP
in sboot/start.S to 0x83000000 (example). This allows
for a huge stack in SDRAM. Note that this can only be
done once we jump to sboot. iboot/eboot should still
use a sp in SRAM since SDRAM will not be initialized
at boot up.

Change-Id: I861ff95bdc2b0e5838c33fe306b73d5219073c60
Signed-off-by: Regis Goupil <x0176713@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>