omap5: clocks: Add support for ES2.0
Some of the L4PER clocks have been moved from the
L4PER_CM_CORE to the CORE_CM_CORE domain.
The original framework didn't allow for setting
the clocks based on ES rev. This patch helps adapt
to checking for the ES rev and sets up the clocks
accordingly.
EMIF1 and L4PER domains need to be explictly enabled.

Change-Id: I59cf0a25420f422fb403b8d4789bd8c3734cdb1a
Signed-off-by: Christina Warren <cawarren@ti.com>