usbboot: Add logs and an error check for HS devices
When user doesn't build with MSHIELD= option, the
HS iboot is not embedded into the usbboot binary.

In the case where the user is trying to flash a HS
device but doesn't provide a signed iboot.ift or
doesn't embed the HS iboot into the usbboot binary,
usbboot will push an embedded GP iboot binary to the
target resulting in a failure with no indication to
the user on what happened. This patch fixes that.

Change-Id: Ifb9879db063e61666c7381e69e9e0e1ebe434e6b
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>
Signed-off-by: Christina Warren <cawarren@ti.com>